Avatar Image in member's home directory
I am just proposing a change, with which, if there is a home directory defined for a member. Then the avatar image is saved in his/her home directory. This change affects only frontend `members` not `users`.  Since there can be a few thousand registered members or more, scanning through so many files may have a performance impact (but not sure). 

What do you think?